# src/main/resources/api.yaml
openapi: 3.0.1
info:
  description: App conjunto integration
  version: 1.0.0
  license:
    name: Licencia Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
  title: Conjunto app Service API
servers:
  - url: https://#uri/api/1.0.0
...
paths:
...
  /web/unidades:
    get:
      tags:
        - web-unidades
      operationId: unidades
      summary: Obtiene una lista de todas las unidades
      parameters:
        - in: query
          name: page
          required: false
          description: Número de registros a devolver por página (paginación).
          schema:
            type: integer
            default: 10
        - in: query
          name: size
          required: false
          description: Número de registros a omitir antes de empezar a devolver los resultados (paginación).
          schema:
            type: integer
            default: 0
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnidadesConjunto"
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idUnidad}/residentes:
    get:
      tags:
        - web-unidades
      operationId: unidadResidentes
      summary: Obtiene una lista de todos los residentes asociados a una unidad
      parameters:
        - name: idUnidad
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          description: Número de registros a devolver por página (paginación).
          schema:
            type: integer
            default: 10
        - in: query
          name: size
          required: false
          description: Número de registros a omitir antes de empezar a devolver los resultados (paginación).
          schema:
            type: integer
            default: 0
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentesUnidad"
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idUnidad}/conceptos:
    get:
      tags:
        - web-unidades
      operationId: unidadeConceptos
      summary: Obtiene una lista de todos los conceptos asociados a una unidad
      parameters:
        - name: idUnidad
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          description: Número de registros a devolver por página (paginación).
          schema:
            type: integer
            default: 10
        - in: query
          name: size
          required: false
          description: Número de registros a omitir antes de empezar a devolver los resultados (paginación).
          schema:
            type: integer
            default: 0
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnidadConceptos"
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idUnidad}/cuentas-cobro:
    get:
      tags:
        - web-unidades
      operationId: unidadCuentasCobro
      summary: Obtiene una lista de todos las cuentas de cobro asociadas a la unidad
      parameters:
        - name: idUnidad
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          description: Número de registros a devolver por página (paginación).
          schema:
            type: integer
            default: 10
        - in: query
          name: size
          required: false
          description: Número de registros a omitir antes de empezar a devolver los resultados (paginación).
          schema:
            type: integer
            default: 0
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnidadCuentasCobro"
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idUnidad}/concepto/{idConcepto}:
    put:
      tags:
        - web-unidades
      operationId: unidadConcepto
      summary: Asocia un concepto con la unidad
      parameters:
        - name: idUnidad
          in: path
          required: true
          schema:
            type: string
        - name: idConcepto
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idUnidad}/residente/{idResidente}:
    put:
      tags:
        - web-unidades
      operationId: unidadResidente
      summary: Asocia un residente con la unidad
      parameters:
        - name: idUnidad
          in: path
          required: true
          schema:
            type: string
        - name: idResidente
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/unidad/{idResidenteUnidad}:
    delete:
      tags:
        - web-unidades
      operationId: eliminaUnidadResidente
      summary: Elimina la relación de unidad con residente
      parameters:
        - name: idResidenteUnidad
          in: path
          required: true
          schema:
            type: string
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      responses:
        '200':
          description: Lista de unidades del residente
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
...
  /web/cuenta-cobro:
    post:
      tags:
        - web-cuenta-cobro
      operationId: crearCuentaCobro
      summary: Registra cuentas de cobro para todas las unidades
      parameters:
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrarCuentaCobro'
      responses:
        '200':
          description: Lista de unidades del residente
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
    get:
      tags:
        - web-cuenta-cobro
      operationId: CuentaCobro
      summary: Consulta todas las cuentas de cobro
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
        - name: size
          in: query
          required: true
          schema:
            type: integer
        - name: idUnidad
          in: query
          required: false
          schema:
            type: string
        - name: estado
          in: query
          required: false
          schema:
            type: string
        - name: fechaInicio
          in: query
          required: false
          schema:
            type: string
        - name: fechaFin
          in: query
          required: false
          schema:
            type: string
        - name: c
          in: query
          required: false
          schema:
            type: string

      responses:
        '200':
          description: Lista de unidades del residente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuentasCobro"
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/cuenta-cobro/unidad:
    post:
      tags:
        - web-cuenta-cobro
      operationId: cuentaCobroUnidad
      summary: Registra una cuenta de cobro para una unidad
      parameters:
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrarCuentaCobroUnidad'
      responses:
        '200':
          description: Lista de unidades del residente
        '400':
          description: Parámetros de solicitud inválidos
        '404':
          description: No se encontraron solicitudes
  /web/cuenta-cobro/pagos/manual:
    post:
      tags:
        - web-cuenta-cobro
      operationId: pagoManual
      summary: Marca un pago manual realizado por un residente (por transferencia, efectivo, etc.)
      description: Permite al administrador registrar un pago manual para una factura (por ejemplo, pagos realizados por transferencia o en efectivo).
      parameters:
        - in: query
          name: c
          required: true
          description: Id del conjunto
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagoManual'
      responses:
        '200':
          description: Pago manual registrado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID del pago registrado
                    example: 456
                  factura_id:
                    type: integer
                    description: ID de la factura asociada al pago
                    example: 345
                  monto_pagado:
                    type: number
                    format: float
                    description: Monto pagado
                    example: 150000.00
                  metodo_pago:
                    type: string
                    enum: ['transferencia', 'efectivo']
                    description: Método de pago utilizado
                    example: "transferencia"
                  fecha_pago:
                    type: string
                    format: date-time
                    description: Fecha en que se registró el pago
                    example: "2025-01-30T14:30:00Z"
        '400':
          description: Error en los datos proporcionados (por ejemplo, factura no encontrada)
        '404':
          description: Factura no encontrada
...
components:
  schemas:
...
    Concepto:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
          example: "Cuota de administración"
        valor:
          type: number
          format: double
          example: 150000
        fechaRegistro:
          type: string
...
    Unidad:
      type: object
      properties:
        id:
          type: string
        interior:
          type: string
        apto:
          type: string
        estado:
          type: string
    UnidadesConjunto:
      type: object
      properties:
        unidades:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/Unidad'
        pages:
          type: object
          $ref: '#/components/schemas/Pages'
    ResidentesUnidad:
      type: object
      properties:
        residentes:
          type: array
          items:
            $ref: '#/components/schemas/Residente'
        pages:
          type: object
          $ref: '#/components/schemas/Pages'
    Residente:
      type: object
      properties:
        idResidenteUnidad:
          type: string
        nombre:
          type: string
        email:
          type: string
        telefono:
          type: string
        tipo:
          type: string
        estado:
          type: string
        fechaRegistro:
          type: string
    UnidadConceptos:
      type: object
      properties:
        conceptos:
          type: array
          items:
            $ref: '#/components/schemas/Concepto'
        pages:
          type: object
          $ref: '#/components/schemas/Pages'
    UnidadCuentasCobro:
      type: object
      properties:
        cuentasCobro:
          type: array
          items:
            $ref: '#/components/schemas/CuentaCobro'
        pages:
          type: object
          $ref: '#/components/schemas/Pages'
    CuentasCobro:
      type: object
      properties:
        cuentasCobro:
          type: array
          items:
            $ref: '#/components/schemas/CuentaCobro'
        pages:
          type: object
          $ref: '#/components/schemas/Pages'
    CuentaCobro:
      type: object
      properties:
        id:
          type: string
        estado:
          type: string
        fechaInicio:
          type: string
        fechaFin:
          type: string
        total:
          type: number
          format: double
        conceptos:
          type: array
          items:
            $ref: '#/components/schemas/Concepto'
        fechaRegistro:
          type: string
        unidad:
          type: object
          $ref: '#/components/schemas/Unidad'
    RegistrarCuentaCobro:
      type: object
      properties:
        fechaInicio:
          type: number
        fechaFin:
          type: number
    RegistrarCuentaCobroUnidad:
      type: object
      properties:
        fechaInicio:
          type: number
        fechaFin:
          type: number
        idConcepto:
          type: number
        idUnidad:
          type: number
    PagoManual:
      type: object
      properties:
        idCuenta:
          type: number
        montoPagado:
          type: number
          format: float
          description: Monto total pagado
          example: 150000.00
        metodoPago:
          type: string
          enum: ['transferencia', 'efectivo']
          description: Método de pago utilizado (puede ser transferencia bancaria o efectivo)
          example: "transferencia"
        fechaPago:
          type: integer
          description: Fecha en la que se realizó el pago
    Pages:
      type: object
      properties:
        size:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        nextPage:
          type: integer
          format: int32